requirements.txt
backports.tarfile>=1.2.0
bcrypt>=4.3.0
bidict>=0.23.1
blinker>=1.9.0
certifi>=2025.7.14
cffi>=1.17.1
charset-normalizer>=3.4.2
click>=8.2.1
colorama>=0.4.6
contourpy>=1.3.2
cryptography>=45.0.5
cycler>=0.12.1
diff-match-patch>=20241021
docutils>=0.21.2
et_xmlfile>=2.0.0
Flask>=3.1.1
Flask-SocketIO>=5.5.1
Flask-SQLAlchemy>=3.1.1
fonttools>=4.59.0
func_timeout>=4.3.5
future>=1.0.0
greenlet>=3.2.3
h11>=0.16.0
id>=1.5.0
idna>=3.10
igraph>=0.10.8
importlib_metadata>=8.7.0
itsdangerous>=2.2.0
jaraco.classes>=3.4.0
jaraco.context>=6.0.1
jaraco.functools>=4.2.1
Jinja2>=3.1.6
junos-eznc>=2.7.4
keyring>=25.6.0
kiwisolver>=1.4.8
logicmonitor-sdk>=3.0.224
lxml>=6.0.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.10.3
mdurl>=0.1.2
more-itertools>=10.7.0
n2g>=0.3.3
napalm>=5.0.0
ncclient>=0.6.15
netaddr>=1.3.0
netmiko>=4.6.0
netutils>=1.14.1
networkx>=3.2
nh3>=0.3.0
ntc_templates>=7.9.0
numpy>=2.3.1
openpyxl>=3.1.5
packaging>=25.0
paramiko>=3.5.1
pillow>=11.3.0
pkginfo>=1.12.1.2
psutil>=7.0.0
pycparser>=2.22
pyeapi>=1.0.4
Pygments>=2.19.2
PyNaCl>=1.5.0
pynetbox>=7.5.0
pyparsing>=3.2.3
PyQt6>=6.9.1
PyQt6-Charts>=6.9.0
PyQt6-Charts-Qt6>=6.9.1
PyQt6-Qt6>=6.9.1
PyQt6-WebEngine>=6.9.0
PyQt6-WebEngine-Qt6>=6.9.1
PyQt6_sip>=13.10.2
pyserial>=3.5
pysnmp>=7.1
python-dateutil>=2.9.0.post0
python-engineio>=4.12.2
python-igraph>=0.10.8
python-socketio>=5.13.0
pywin32-ctypes>=0.2.3
PyYAML>=6.0.2
qasync>=0.27.1
readme_renderer>=44.0
requests>=2.32.4
requests-toolbelt>=1.0.0
rfc3986>=2.0.0
rich>=14.0.0
ruamel.yaml>=0.18.14
ruamel.yaml.clib>=0.2.12
scp>=0.15.0
setuptools>=80.9.0
simple-websocket>=1.1.0
six>=1.17.0
SQLAlchemy>=2.0.41
textfsm>=1.1.3
texttable>=1.7.0
transitions>=0.9.3
ttp>=0.9.5
ttp-templates>=0.3.7
twine>=6.1.0
typing_extensions>=4.14.1
urllib3>=2.5.0
Werkzeug>=3.1.3
wsproto>=1.2.0
yamlordereddictloader>=0.4.2
zipp>=3.23.0
